Separate BasicList into own file
