+++ lib/index.ts
@@ -1,2 +1,3 @@
+export * from './basic-list';
 export * from './dual-list.component';
 export * from './angular-dual-listbox.module';
